Break out cluster into it's own module.
