+++ templates/index.dtl
@@ -21,8 +21,9 @@
     <script src="/admin/ui/js/actions.js"></script>
 
     <script src="/admin/ui/js/app.js"></script>
-    <script src="/admin/ui/js/ring.js"></script>
     <script src="/admin/ui/js/snapshot.js"></script>
+    <script src="/admin/ui/js/cluster.js"></script>
+    <script src="/admin/ui/js/ring.js"></script>
 
     <script>minispade.require('app');</script>
 </body>
